+++ install.ps1
@@ -4,7 +4,7 @@ param(
 )
 
 $ErrorActionPreference = 'Stop'
-$supportedVersions = @('0.1.2-alpha.3')
+$supportedVersions = @('0.1.7-rc.1')
 $dshCommand = Get-Command dsh -ErrorAction Stop
 $actualVersion = (& $dshCommand.Source --version).Trim()